split DataSpec
